html-website-templates builder advance

+++ src/server/client-build.js
@@ -40,9 +40,22 @@ const fullBuild = async ({
 }) => {
   logger.warn('Full build', rootClientPath);
 
-  fs.removeSync(rootClientPath);
-
   buildAcmeChallengePath(acmeChallengeFullPath);
+
+  if (publicClientId.startsWith('html-website-templates')) {
+    if (!fs.existsSync(`/dd/html-website-templates/`))
+      shellExec(`cd /dd && git clone https://github.com/designmodo/html-website-templates.git`);
+    if (!fs.existsSync(`${rootClientPath}/index.php`)) {
+      // ->
+      fs.copySync(`/dd/html-website-templates/${publicClientId.split('-publicClientId-')[1]}`, rootClientPath);
+      shellExec(`cd ${rootClientPath} && git init && git add . && git commit -m "Base template implementation"`);
+      // git remote add origin git@github.com:<username>/<repo>.git
+      fs.writeFileSync(`${rootClientPath}/.git/.htaccess`, `Deny from all`, 'utf8');
+    }
+    return;
+  }
+
+  fs.removeSync(rootClientPath);
 
   if (fs.existsSync(`./src/client/public/${publicClientId}`)) {
     if (iconsBuild) {
@@ -64,15 +77,6 @@ const fullBuild = async ({
           },
         } */,
     );
-  } else if (publicClientId.startsWith('html-website-templates')) {
-    if (!fs.existsSync(`/dd/html-website-templates/`))
-      shellExec(`cd /dd && git clone https://github.com/designmodo/html-website-templates.git`);
-    if (!fs.existsSync(`${rootClientPath}/index.php`)) {
-      fs.copySync(`/dd/html-website-templates/${publicClientId.split('-publicClientId-')[1]}`, rootClientPath);
-      shellExec(`cd ${rootClientPath} && git init && git add . && git commit -m "Base template implementation"`);
-      // git remote add origin git@github.com:<username>/<repo>.git
-      fs.writeFileSync(`${rootClientPath}/.git/.htaccess`, `Deny from all`, 'utf8');
-    }
   } else if (fs.existsSync(`./engine-private/src/client/public/${publicClientId}`)) {
     switch (publicClientId) {
       case 'mysql_test':
